requirements.txt
llama-cpp-python
numpy
neat-python
graphviz
matplotlib
gym